+++ servers/libvirt/libvirt.php
@@ -57,50 +57,9 @@ function libvirt_MetaData()
  */
 function libvirt_ConfigOptions()
 {
-    return array(
-        // a text field type allows for single line text input
-        'Text Field' => array(
-            'Type' => 'text',
-            'Size' => '25',
-            'Default' => '1024',
-            'Description' => 'Enter in megabytes',
-        ),
-        // a password field type allows for masked text input
-        'Password Field' => array(
-            'Type' => 'password',
-            'Size' => '25',
-            'Default' => '',
-            'Description' => 'Enter secret value here',
-        ),
-        // the yesno field type displays a single checkbox option
-        'Checkbox Field' => array(
-            'Type' => 'yesno',
-            'Description' => 'Tick to enable',
-        ),
-        // the dropdown field type renders a select menu of options
-        'Dropdown Field' => array(
-            'Type' => 'dropdown',
-            'Options' => array(
-                'option1' => 'Display Value 1',
-                'option2' => 'Second Option',
-                'option3' => 'Another Option',
-            ),
-            'Description' => 'Choose one',
-        ),
-        // the radio field type displays a series of radio button options
-        'Radio Field' => array(
-            'Type' => 'radio',
-            'Options' => 'First Option,Second Option,Third Option',
-            'Description' => 'Choose your option!',
-        ),
-        // the textarea field type allows for multi-line text input
-        'Textarea Field' => array(
-            'Type' => 'textarea',
-            'Rows' => '3',
-            'Cols' => '60',
-            'Description' => 'Freeform multi-line text input field',
-        ),
-    );
+    return [
+        
+    ];
 }
 
 /**
@@ -369,7 +328,7 @@ function libvirt_TestConnection(array $params)
 {        
     try {
         // Call the service's connection test function.       
-        $api = new Libvirt($params['serverusername'] . '@' . $params['serverip']);
+        $api = new Libvirt($params['serverusername'], $params['serverip']);
         
         if ($api->testConnection()) {
             $success = true;
@@ -405,12 +364,24 @@ function libvirt_TestConnection(array $params)
  *
  * @return array
  */
+// function libvirt_AdminCustomButtonArray()
+// {
+//     return array(
+//         "Reboot Server" => "buttonOneFunction",
+//         "Shutdown Server" => "buttonTwoFunction",
+//     );
+// }
+
 function libvirt_AdminCustomButtonArray()
 {
-    return array(
-        "Reboot Server" => "buttonOneFunction",
-        "Shutdown Server" => "buttonTwoFunction",
-    );
+    global $_LANG;
+
+    return [
+        $_LANG["libvirt_start"] => "start",
+        "Shutdown Server" => "shutdown",
+        "Reboot Server" => "reboot",
+        "Reset Server" => "reset",
+    ];
 }
 
 /**
@@ -426,13 +397,26 @@ function libvirt_AdminCustomButtonArray()
  */
 function libvirt_ClientAreaCustomButtonArray()
 {
-    return array(
+    global $_LANG;
+
+    return [
         "Start Server" => "start",
         "Shutdown Server" => "shutdown",
         "Reboot Server" => "reboot",
         "Reset Server" => "reset",
-    );
+    ];
 }
+
+// SolusVM example
+// function libvirt_ClientAreaCustomButtonArray() {
+//     global $_LANG;
+
+//     return array(
+//         $_LANG["solusvmpro_reboot"]   => "reboot",
+//         $_LANG["solusvmpro_shutdown"] => "shutdown",
+//         $_LANG["solusvmpro_boot"]     => "boot",
+//     );
+// }
 
 /**
  * Custom function for performing an additional action.